refactored models
