--- locksmith/src/block.js
@@ -1,4 +1,0 @@
-const models = require('../models')
-
-models.Block.removeAttribute('id')
-module.exports = models.Block
