--- locksmith/src/lock.js
@@ -1,4 +1,0 @@
-const models = require('../models')
-
-models.Lock.removeAttribute('id')
-module.exports = models.Lock
