--- locksmith/src/transaction.js
@@ -1,4 +1,0 @@
-const models = require('../models')
-
-models.Transaction.removeAttribute('id')
-module.exports = models.Transaction
